Added more empty objects and a script
I added a script for the movement, its kind of broken right now because I need a ground already set up so the object of characters can move around. There are also like a ball I added and is just for testing. Please don't delete the actual objects, just replace or change the needed componets.